// src/components/Spotify/types.ts
export interface ExternalUrls {
  spotify: string;
}

export interface Artist {
  external_urls: ExternalUrls;
  href: string;
  id: string;
  name: string;
  type: string;
  uri: string;
}

export interface Image {
  height: number;
  url: string;
  width: number;
}

export interface SongData {
  name: string;
  artists: Artist[];
  image: Image;
  date: string;
  albumName: string;
}